Se solucionaron los errores al implementar la nueva base de datos, se agrgaron los put en los controladores.

+++ src/main/java/com/zeal/repository/MensajesRepository.java
@@ -1,11 +1,14 @@
 package com.zeal.repository;
 
 import com.zeal.model.MensajesModel;
+
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MensajesRepository extends JpaRepository<MensajesModel, Long> {
+public interface MensajesRepository extends JpaRepository<MensajesModel, Integer> {
